Make ui.js
