+++ index.html
@@ -22,7 +22,7 @@
               <span id="path-directory-inner"></span>
             </span>
             <h5 class="nav-group-title">Files</h5>
-            <nav id="directory-link"></nav>
+            <nav id="directory-tree"></nav>
           </nav>
         </div>
