+++ lib/sy.js
@@ -1,5 +1,4 @@
-const path = require('path')
-const directoryLink = document.getElementById('directory-link')
+'use strict'
 
 module.exports.isDirectory = function(filePath) {
   const fs = require('fs');
@@ -14,17 +13,4 @@ module.exports.isDirectory = function(filePath) {
 module.exports.openDirectory = function(path) {
   const { spawn } = require('child_process');
   spawn('open', [path]);
-}
-
-module.exports.getCurrent = function() {
-  var current = document.getElementById('directory-current-page')
-  if (current) {
-    return current
-  }
-
-  current = directoryLink.firstChild
-  if (!current) return
-
-  current.id = 'directory-current-page'
-  return current
 }
